[BE] 스케줄 수정 및 삭제할 때, 그 스케줄에 예약이 되어 있으면 SqlException 발생 문제 해결 (#199)
* feat: service에 일정 삭제 검증 로직 추가

* test: 일정 수정 불가능한 상황 컨트롤러 테스트 추가

* refactor: return type 수정

* style: 메서드 이름 변경

* refactor: 가독성을 위해 메서드 분리

* refactor: for문 사용해 성능 향상

+++ back/src/main/java/com/woowacourse/teatime/service/ScheduleService.java
@@ -6,6 +6,7 @@ import com.woowacourse.teatime.controller.dto.response.ScheduleFindResponse;
 import com.woowacourse.teatime.domain.Coach;
 import com.woowacourse.teatime.domain.Schedule;
 import com.woowacourse.teatime.exception.NotFoundCoachException;
+import com.woowacourse.teatime.exception.UnableToUpdateSchedule;
 import com.woowacourse.teatime.repository.CoachRepository;
 import com.woowacourse.teatime.repository.ScheduleRepository;
 import com.woowacourse.teatime.util.Date;
@@ -34,19 +35,13 @@ public class ScheduleService {
 
     @Transactional(readOnly = true)
     public List<ScheduleFindResponse> find(Long coachId, ScheduleFindRequest request) {
-        validateCoachId(coachId);
+        findCoach(coachId);
 
         LocalDateTime start = Date.findFirstDay(request.getYear(), request.getMonth());
         LocalDateTime end = Date.findLastDay(request.getYear(), request.getMonth());
         List<Schedule> schedules
                 = scheduleRepository.findByCoachIdAndLocalDateTimeBetweenOrderByLocalDateTime(coachId, start, end);
         return ScheduleFindResponse.from(schedules);
-    }
-
-    private void validateCoachId(Long id) {
-        if (!coachRepository.existsById(id)) {
-            throw new NotFoundCoachException();
-        }
     }
 
     public void update(Long coachId, ScheduleUpdateRequest request) {
@@ -58,14 +53,38 @@ public class ScheduleService {
         LocalDate date = request.getDate();
         LocalDateTime start = Date.findFirstTime(date);
         LocalDateTime end = Date.findLastTime(date);
+        validateDeletable(coachId, request, start, end);
+
         scheduleRepository.deleteAllByCoachIdAndLocalDateTimeBetween(coachId, start, end);
     }
 
+    private void validateDeletable(Long coachId, ScheduleUpdateRequest request, LocalDateTime start,
+                                   LocalDateTime end) {
+        Coach coach = findCoach(coachId);
+        List<Schedule> newSchedules = toSchedules(request, coach);
+        List<Schedule> oldSchedules = scheduleRepository
+                .findByCoachIdAndLocalDateTimeBetweenOrderByLocalDateTime(coachId, start, end);
+
+        for (Schedule schedule : oldSchedules) {
+            ValidateIsReserved(newSchedules, schedule);
+        }
+    }
+
+    private void ValidateIsReserved(List<Schedule> newSchedules, Schedule schedule) {
+        if (!(newSchedules.contains(schedule) || schedule.isPossible())) {
+            throw new UnableToUpdateSchedule();
+        }
+    }
+
     private void saveAllByCoachAndDate(Long coachId, ScheduleUpdateRequest request) {
-        Coach coach = coachRepository.findById(coachId)
-                .orElseThrow(NotFoundCoachException::new);
+        Coach coach = findCoach(coachId);
         List<Schedule> schedules = toSchedules(request, coach);
         scheduleRepository.saveAll(schedules);
+    }
+
+    private Coach findCoach(Long id) {
+        return coachRepository.findById(id)
+                .orElseThrow(NotFoundCoachException::new);
     }
 
     private List<Schedule> toSchedules(ScheduleUpdateRequest request, Coach coach) {
@@ -73,5 +92,4 @@ public class ScheduleService {
                 .map(schedule -> new Schedule(coach, schedule))
                 .collect(Collectors.toList());
     }
-
 }
